requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==2.3.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.1.4
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.12.2
urllib3==2.0.7
Werkzeug==2.3.7